requirements.txt
clustaar.schemas==0.1.3
clustaar.webhook==0.1.2
falcon==1.4.1
gunicorn==19.9.0
jmespath==0.9.3
lupin==1.0.2
python-mimeparse==1.6.0
six==1.11.0
tinydb==3.10.0
ujson==1.35
